refactor(settings): CustomAcpAgent 目录名改为 snake_case
